requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
textblob
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pytest
pytest-asyncio